Spring core was integrated into the project

+++ src/main/java/com/example/elective/dao/interfaces/CourseDAO.java
@@ -24,6 +24,6 @@ public interface CourseDAO extends DAO<Course> {
 
   List<Course> getAvailableForStudent(int studentId);
 
-  int getCount(int teacherId);
+  long getCount(int teacherId);
 
 }
