+++ src/main/java/com/example/elective/selection/CoursePagination.java
@@ -7,7 +7,7 @@ package com.example.elective.selection;
 
 public class CoursePagination extends Pagination {
 
-  public CoursePagination(int page, int totalItems) {
+  public CoursePagination(int page, long totalItems) {
     super(page, 1, totalItems);
   }
